+++ IB.Api.Client/Client/Model/AccountUpdate.cs
@@ -24,19 +24,19 @@ namespace IB.Api.Client.Model
         public string AccountType { get; set; }
 
         [JsonProperty("availableFunds")]
-        public double AvailableFunds { get; set; }
+        public decimal AvailableFunds { get; set; }
 
         [JsonProperty("buyingPower")]
-        public double BuyingPower { get; set; }
+        public decimal BuyingPower { get; set; }
 
         [JsonProperty("cashGbp")]
-        public double CashGbp { get; set; }
+        public decimal CashGbp { get; set; }
 
         [JsonProperty("cashUsd")]
-        public double CashUsd { get; set; }
+        public decimal CashUsd { get; set; }
 
         [JsonProperty("excessLiquidity")]
-        public double ExcessLiquidity { get; set; }
+        public decimal ExcessLiquidity { get; set; }
 
         [JsonProperty("side")]
         public string Side { get; set; }
@@ -45,20 +45,20 @@ namespace IB.Api.Client.Model
         public string Pnl { get; set; }
 
         [JsonProperty("cashUsdControl")]
-        public double CashUsdControl { get; set; }
+        public decimal CashUsdControl { get; set; }
 
         [JsonProperty("cashUsdNet")]
-        public double CashUsdNet { get; set; }
+        public decimal CashUsdNet { get; set; }
 
         public void SetValue(string key, string value, string currency)
         {
             AccountCode = key == "AccountCode" ? value : AccountCode;
             AccountType = key == "AccountType" ? value : AccountType;
-            AvailableFunds = key == "AvailableFunds" ? double.Parse(value) : AvailableFunds;
-            BuyingPower = key == "BuyingPower" ? double.Parse(value) : BuyingPower;
-            CashGbp = key == "CashBalance" && currency == "GBP" ? double.Parse(value) : CashGbp;
-            CashUsd = key == "CashBalance" && currency == "USD" ? double.Parse(value) : CashUsd;
-            ExcessLiquidity = key == "ExcessLiquidity" ? double.Parse(value) : ExcessLiquidity;
+            AvailableFunds = key == "AvailableFunds" ? decimal.Parse(value) : AvailableFunds;
+            BuyingPower = key == "BuyingPower" ? decimal.Parse(value) : BuyingPower;
+            CashGbp = key == "CashBalance" && currency == "GBP" ? decimal.Parse(value) : CashGbp;
+            CashUsd = key == "CashBalance" && currency == "USD" ? decimal.Parse(value) : CashUsd;
+            ExcessLiquidity = key == "ExcessLiquidity" ? decimal.Parse(value) : ExcessLiquidity;
         }
     }
 }
